+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to Valuify will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - TBC
+
+## Added
+
+- Added a new `Ignore` attribute that allows engineers to mark properties that should be disregarded from equality checks (#14).
+- Added a new `VALFY04` analyzer which will warn engineers whenever a property is annotated with the `Ignore` attribute but the `Valuify` attribute is missing from the `class`.
+
 ## [1.2.2] - 2025-01-08
 
 ## Fixed